+++ paper/paper.bib
@@ -10,7 +10,7 @@ doi = {10.1137/141000671},
 }
 
 @article{christ2023plots,
-  doi = {https://doi.org/10.5334/jors.431},
+  doi = {10.5334/jors.431},
   author = {Christ, Simon and Schwabeneder, Daniel and Rackauckas, Christopher and Borregaard, Michael Krabbe and Breloff, Thomas},
   title = {Plots.jl -- a user extendable plotting API for the julia programming language},
   publisher = {Journal of Open Research Software},
@@ -34,7 +34,7 @@ doi = {10.1137/141000671},
   author = {Mildenberger, T. and Rozenholc, Y. and Zasada, D.},
   year = {2019},
   note = {R package version 0.0-25},
-  url = {https://CRAN.R-project.org/package=histogram},
+  doi = {10.32614/cran.package.histogram}
 }
 
 @article{sturges1926choice,
