### 2025-07-21 d28e681
Started work on a new, easier to maintain/extend API.
