+++ src/regular/regular_rules.jl
@@ -1,0 +1,166 @@
+struct RRH <: AbstractRegularRule
+    a::Union{Real, Function}
+    logprior::Function
+    maxbins::Union{Int, Symbol}
+end
+
+"""
+    RRH(; a::Union{Real, Function}, logprior::Function, maxbins::Union{Int, Symbol}=:default)
+    Knuth(; maxbins::Union{Int, Symbol}=:default)
+
+The random regular histogram criterion for regular histograms.
+
+The number ``k`` of bins is chosen as the maximizer of the marginal log-posterior,
+```math
+   n\\log (k) + \\sum_{j=1}^k \\big\\{\\log \\Gamma(a_j + N_j) - \\log \\Gamma(a_j)\\big\\} + \\log p_n(k).
+```
+Here ``p_n(k)`` is the prior distribution on the number ``k`` of bins, which can be controlled by
+supplying a function to the `logprior` keyword argument. 
+The default value is ``p_n(k) \\propto 1``. Here, ``a_j = a/k``, for a scalar ``a > 0``,
+possibly depending on ``k``. The value of ``a`` can be set by supplying a fixed, positive scalar or
+a function ``a(k)`` to the keyword argument `a`. The default value is `a=5.0` for `RRH()`.
+
+The rule `Knuth()` is a special case of the `RRH` criterion, which corresponds to the particular
+choices ``a_j = 0.5`` and ``p_n(k)\\propto 1``.
+
+# Keyword arguments
+- `a`: Specifies Dirichlet concentration parameter in the Bayesian histogram model. Can either be a fixed positive number or a function computing aₖ for different values of k. Defaults to `5.0` if not supplied.
+- `logprior`: Unnormalized logprior distribution on the number ``k`` of bins. Defaults to a uniform prior, e.g. `logprior(k) = 0` for all `k`.
+- `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)`.
+
+# References
+The `Knuth` criterion for histograms was proposed by [Knuth (2019)](https://doi.org/10.1016/j.dsp.2019.102581). The random regular histogram criterion is a generalization.
+"""
+RRH, Knuth
+
+RRH(; a::Union{Real, Function}=5.0, logprior::Function=k->0.0, maxbins::Union{Int, Symbol}=:default) = RRH(a, logprior, maxbins)
+
+function fit(::Type{AutomaticHistogram}, x::AbstractVector{<:Real}, rule::RRH; support::Tuple{Real,Real}=(-Inf,Inf), closed::Symbol=:right)
+    if !(closed in [:right, :left]) # if supplied symbol is nonsense, just use default
+        throw(ArgumentError("The supplied value of the closed keyword, :$closed, is invalid. Valid values are :left or :right."))
+    end
+    if maxbins < 1             # maximal number of bins must be positive
+        throw(DomainError("Maximal number of bins must be positive."))
+    end
+    xmin, xmax = check_support(x, support[1], support[2])
+    n = length(x)
+    aₖ = Vector{Float64}(undef, k_max)
+    if !isa(a, Function) # create constant function if typeof(a) <: Real
+        if a ≤ 0.0
+            throw(DomainError("Supplied value of a must be positive."))
+        else 
+            aₖ[1:end] .= a
+        end
+    else 
+        for k = 1:k_max
+            a_curr = a(k)
+            aₖ[k] = a_curr
+        end
+        if min(a_curr) ≤ 0.0
[…]
