### 2025-07-21 f01b767
Finished new API for regular histograms
