+++ src/AutoHist.jl
@@ -4,6 +4,7 @@ export AutomaticHistogram
 export fit, convert, loglikelihood, logmarginallikelihood, minimum, maximum, extrema, peaks, pdf, cdf, insupport, length, distance
 export histogram_regular, histogram_irregular
 export DP
+export RRH, Knuth, AIC, BIC, BR, MDL, NML_R, L2CV_R, KLCV_R, Sturges, FD, Scott, Wand
 
 using StatsBase, Base.Threads, Distributions
 import StatsBase: modes
@@ -12,6 +13,12 @@ import SpecialFunctions: loggamma, logabsbinomial, gamma
 import FFTW: fft, ifft
 import StatsAPI: fit, loglikelihood
 import Distributions: pdf, insupport
+
+abstract type AbstractRule end
+
+abstract type AbstractRegularRule <: AbstractRule end
+
+abstract type AbstractIrregularRule <: AbstractRule end
 
 include("AutomaticHistogram.jl")
 include("utils.jl")
@@ -28,12 +35,6 @@ include(joinpath("irregular", "compute_bounds.jl"))
 include(joinpath("regular", "regular_histogram.jl"))
 include(joinpath("regular", "regular_rules.jl"))
 include(joinpath("irregular", "irregular_histogram.jl"))
-
-abstract type AbstractRule end
-
-abstract type AbstractRegularRule end
-
-abstract type AbstractIrregularRule end
 
 function __init__()
     return nothing
