+++ src/regular/regular_rules.jl
@@ -1,3 +1,5 @@
+# ------------------------------
+# Random regular histogram 
 struct RRH <: AbstractRegularRule
     a::Union{Real, Function}
     logprior::Function
@@ -35,28 +37,22 @@ RRH, Knuth
 
 RRH(; a::Union{Real, Function}=5.0, logprior::Function=k->0.0, maxbins::Union{Int, Symbol}=:default) = RRH(a, logprior, maxbins)
 
-function fit(::Type{AutomaticHistogram}, x::AbstractVector{<:Real}, rule::RRH; support::Tuple{Real,Real}=(-Inf,Inf), closed::Symbol=:right)
-    if !(closed in [:right, :left]) # if supplied symbol is nonsense, just use default
-        throw(ArgumentError("The supplied value of the closed keyword, :$closed, is invalid. Valid values are :left or :right."))
-    end
-    if maxbins < 1             # maximal number of bins must be positive
-        throw(DomainError("Maximal number of bins must be positive."))
-    end
-    xmin, xmax = check_support(x, support[1], support[2])
+function fit_autohist(x::AbstractVector{T}, rule::RRH, xmin::T, xmax::T, closed::Symbol) where {T <: Real}
     n = length(x)
-    aₖ = Vector{Float64}(undef, k_max)
-    if !isa(a, Function) # create constant function if typeof(a) <: Real
-        if a ≤ 0.0
+    k_max = get_maxbins_regular(rule.maxbins, n)
+    a_vec = Vector{Float64}(undef, k_max)
+    if !isa(rule.a, Function) # create constant function if typeof(a) <: Real
+        if rule.a ≤ 0.0
             throw(DomainError("Supplied value of a must be positive."))
         else 
-            aₖ[1:end] .= a
+            a_vec[1:end] .= rule.a
         end
     else 
         for k = 1:k_max
-            a_curr = a(k)
-            aₖ[k] = a_curr
+            a_curr = rule.a(k)
+            a_vec[k] = a_curr
         end
-        if min(a_curr) ≤ 0.0
+        if minimum(a_vec) ≤ 0.0
             throw(DomainError("Supplied function a(k) must return strictly positive values."))
         end
     end
@@ -64,21 +60,27 @@ function fit(::Type{AutomaticHistogram}, x::AbstractVector{<:Real}, rule::RRH; s
     z = @. (x - xmin) / (xmax - xmin)         # Scale data to the interval [0,1]
     Threads.@threads for k = 1:k_max
         N = bin_regular(z, 0.0, 1.0, k, closed == :right)
-        @inbounds criterion[k] = logposterior_k(N, k, ones(k)/k, aₖ[k], n, logprior)
+        @inbounds criterion[k] = logposterior_k(N, k, ones(k)/k, a_vec[k], n, rule.logprior)
     end
+    k_opt = argmax(criterion)
     N = bin_irregular_int(x, LinRange(xmin, xmax, k_opt+1), closed == :right)
-    a_opt = aₖ[k_opt]
+    a_opt = a_vec[k_opt]
     dens = k_opt/(xmax-xmin) * (N .+ a_opt/k_opt) / (a_opt + n) # Estimated density
-    h = AutomaticHistogram(LinRange(xmin, xmax, k_opt+1), dens, N, :regular, closed, ifelse(a_opt > 0.0, a_opt, NaN))
+    h = AutomaticHistogram(LinRange(xmin, xmax, k_opt+1), dens, N, :regular, closed, a_opt)
     return h
 end
 
+
+# ------------------------------
+# Knuth's rule
 struct Knuth <: AbstractRegularRule
[…]
