### 2025-07-21 b6d1ea8
Started work on the API change for irregular histograms.
