+++ src/AutomaticHistogram.jl
@@ -136,7 +136,7 @@ end
 
 # New
 """
-    fit(AutomaticHistogram, x::AbstractVector{x<:Real}, rule<:AbstractRule=RIH(); support::Tuple{Real,Real}=(-Inf,Inf), closed::Symbol=:right)
+    fit(AutomaticHistogram, x::AbstractVector{x<:Real}, rule<:AbstractRule; support::Tuple{Real,Real}=(-Inf,Inf), closed::Symbol=:right)
 
 Fit a histogram to a one-dimensional vector `x` with an automatic and data-based selection of the histogram partition.
 
