+++ src/irregular/algorithms.jl
@@ -76,25 +76,25 @@ The optimal partitioning algorithm for constructing an irregular histogram.
 ```jldoctest 
 julia> using AutoHist
 
-julia> x = LinRange(eps(), 1.0-eps(), 5000) .^(1.0/4.0);
+julia> x = LinRange(eps(), 1.0-eps(), 1000) .^(1.0/4.0);
 
-julia> h = fit(AutomaticHistogram, x, L2CV_I(alg = OptPart(greedy=true, gr_maxbins=200)))
+julia> h = fit(AutomaticHistogram, x, L2CV_I(alg = OptPart(greedy=true, gr_maxbins=100)))
 AutomaticHistogram
-breaks: [0.00012207031249977798, 0.17763663029325183, 0.29718725232110504, 0.4022468898607337, 0.4928155429121377, 0.5797614498414855, 0.6667073567708333, 0.7572760098222373, 0.8405991706295289, 0.9239223314368207, 1.0000000000000002]
-density: [0.00662683597412854, 0.057821970706400425, 0.17596277991076312, 0.36279353706969375, 0.6214544825215076, 0.9730458529384184, 1.4481767793920146, 2.0440057561776532, 2.7513848134529346, 3.564842182949155]
-counts: [5, 34, 92, 164, 270, 423, 656, 852, 1147, 1357]
+breaks: [0.00012207031249977798, 0.16676839192708331, 0.2977047874813988, 0.4048345656622024, 0.5119643438430059, 0.6071908133370535, 0.6786106654575893, 0.7619338262648809, 0.8452569870721726, 0.9285801478794643, 1.0000000000000002]
+density: [0.006000732511292875, 0.05346107146424569, 0.17735498311154516, 0.3920478574044684, 0.7035858869490909, 1.0641298986692698, 1.5001831278232214, 2.0762534489073383, 2.7963413502624808, 3.598439262605288]
+counts: [1, 7, 19, 42, 67, 76, 125, 173, 233, 257]
 type: irregular
 closed: right
-a: 5.0
+a: NaN
 
 julia> h = fit(AutomaticHistogram, x, L2CV_I(alg = OptPart(greedy=false)))
 AutomaticHistogram
-breaks: [0.00012207031249977798, 0.17763663029325183, 0.29718725232110504, 0.4022468898607337, 0.4928155429121377, 0.5797614498414855, 0.6667073567708333, 0.7572760098222373, 0.8405991706295289, 0.9202995853147645, 1.0000000000000002]
-density: [0.00662683597412854, 0.057821970706400425, 0.17596277991076312, 0.36279353706969375, 0.6214544825215076, 0.9730458529384184, 1.4481767793920146, 2.0440057561776532, 2.733509595364622, 3.545742066060367]
-counts: [5, 34, 92, 164, 270, 423, 656, 852, 1090, 1414]
+breaks: [0.00012207031249977798, 0.16676839192708331, 0.2977047874813988, 0.4048345656622024, 0.5119643438430059, 0.6071908133370535, 0.6786106654575893, 0.7619338262648809, 0.8452569870721726, 0.9285801478794643, 1.0000000000000002]
+density: [0.006000732511292875, 0.05346107146424569, 0.17735498311154516, 0.3920478574044684, 0.7035858869490909, 1.0641298986692698, 1.5001831278232214, 2.0762534489073383, 2.7963413502624808, 3.598439262605288]
+counts: [1, 7, 19, 42, 67, 76, 125, 173, 233, 257]
 type: irregular
 closed: right
-a: 5.0
+a: NaN
 ```
 """
 function OptPart(; greedy::Bool=true, gr_maxbins::Union{Int, Symbol}=:default)
