### 2025-07-22 6f8bf27
Improved type inference for regular_rules and irregular_rules.
