+++ src/irregular/irregular_rules.jl
@@ -1,11 +1,11 @@
 # ------------------------------
 # Random irregular histogram 
-struct RIH <: AbstractIrregularRule
-    a::Real
+struct RIH{T<:Real, A<:AbstractAlgorithm} <: AbstractIrregularRule
+    a::T
     logprior::Function
     grid::Symbol
     maxbins::Union{Int, Symbol}
-    alg::AbstractAlgorithm
+    alg::A
 end
 
 """
@@ -32,7 +32,7 @@ The default value is ``p_n(k) \\propto 1``. Here, ``a_j = a/k``, for a scalar ``
 - `logprior`: Unnormalized logprior distribution on the number ``k`` of bins. Defaults to a uniform prior, e.g. `logprior(k) = 0` for all `k`.
 - `grid`: Symbol indicating how the finest possible mesh should be constructed. Options are `:data`, which uses each unique data point as a grid point, `:regular` (default) which constructs a fine regular grid, and `:quantile` which constructs the grid based on the sample quantiles.
 - `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)` if `grid` is `regular` or `quantile`. Ignored if `grid=:data`.
-- `alg`: Algorithm used to fit the model. Currently, only `SegNeig()` for this rule. See [`DP`](@ref) for further details.
+- `alg`: Algorithm used to fit the model. Currently, only `SegNeig()` for this rule. See [`SegNeig`](@ref) for further details.
 
 # References
 This approach to irregular histograms first appeared in [Simensen et al. (2025)](https://doi.org/10.48550/ARXIV.2505.22034).
@@ -57,10 +57,10 @@ end
 
 # ------------------------------
 # RMG penA
-struct RMG_penA <: AbstractIrregularRule
+struct RMG_penA{A<:AbstractAlgorithm} <: AbstractIrregularRule
     grid::Symbol
     maxbins::Union{Int, Symbol}
-    alg::AbstractAlgorithm
+    alg::A
 end
 
 """
@@ -78,7 +78,7 @@ Consists of finding the partition ``\\mathcal{I}`` that maximizes a penalized lo
 # Keyword arguments
 - `grid`: Symbol indicating how the finest possible mesh should be constructed. Options are `:data`, which uses each unique data point as a grid point, `:regular` (default) which constructs a fine regular grid, and `:quantile` which constructs the grid based on the sample quantiles.
 - `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)` if `grid` is `regular` or `quantile`. Ignored if `grid=:data`.
-- `alg`: Algorithm used to fit the model. Currently, only `SegNeig()` is supported for this rule. See [`DP`](@ref) for further details.
+- `alg`: Algorithm used to fit the model. Currently, only `SegNeig()` is supported for this rule. See [`SegNeig`](@ref) for further details.
 
 # References
 This approach was suggested by [Rozenholc et al. (2010)](https://doi.org/10.1016/j.csda.2010.04.021).
@@ -91,8 +91,7 @@ function fit_autohist(x::AbstractVector{T}, rule::RMG_penA, xmin::T, xmax::T, cl
     
     bin_edges = @. xmin + (xmax - xmin) * bin_edges_norm
     N = bin_irregular_int(x, bin_edges, closed == :right)
[…]
