### 2025-07-22 b2fb9f1
Tested for errors in greedy algorithm.
