### 2025-07-22 2660745
Fixed bug where binning algorithm would sometimes go out of bounds
