### 2025-07-22 7a1862b
Added new doctests.
