+++ src/AutomaticHistogram.jl
@@ -79,62 +79,7 @@ function Base.show(io::IO, h::AutomaticHistogram)
     println(io, "a: ", h.a)
 end
 
-# Old
-"""
-    fit(AutomaticHistogram, x::AbstractVector{x<:Real}; rule=:default, type=:irregular, kwargs...)
 
-Fit a histogram to a one-dimensional vector `x` with an automatic and data-based selection of the histogram partition.
-
-# Arguments
-- `x`: 1D vector of data for which a histogram is to be constructed.
-
-# Keyword arguments
-- `rule`: The criterion used to determine the optimal number of bins. Default value is `rule=:default` which uses the default rule for the regular or irregular histogram procedure depending on the value of `type`.
-- `type`: Symbol indicating whether the fitted method is a regular and irregular one. The rules `:bayes`, `:l2cv`, `:klcv` and `:nml` are implemented for both regular and irregular histograms, and this keyword specifies whether the regular or irregular version should be used. For other rules, this function automatically infers the type from the `rule` keyword, and misspecifying the rule in this case has no effect. Possible values are `:irregular` (default) and `:regular`.
-- `kwargs`: Additional keyword arguments passed to [`histogram_regular`](@ref) or [`histogram_irregular`](@ref) depending on the specified or inferred type.
-
-# Returns
-- `h`: An object of type [`AutomaticHistogram`](@ref), corresponding to the fitted histogram.
-
-# Examples
-```julia
-julia> x = randn(10^3)
-julia> h1 = fit(AutomaticHistogram, x)                                      # fits an irregular histogram
-julia> h2 = fit(AutomaticHistogram, x; rule=:wand, scalest=:stdev, level=4) # fits a regular histogram
-```
-"""
-#= function fit(::Type{AutomaticHistogram}, x::AbstractVector{<:Real}; rule::Symbol=:default, type::Symbol=:irregular, kwargs...)
-    if rule == :default
-        if type == :irregular
-            rule = :bayes
-        elseif type == :regular
-            rule = :knuth
-        else 
-            throw(ArgumentError("The supplied type, :$type, is not supported and the supplied rule=:default could as such not be inferred. Valid types are :irregular and :regular."))
-        end
-    elseif rule in [:aic, :bic, :br, :mdl, :sturges, :fd, :scott, :wand]
-        type = :regular
-    elseif rule in [:pena, :penb, :penr]
-        type = :irregular
-    elseif rule in [:bayes, :l2cv, :klcv, :nml]
-        if !(type in [:regular, :irregular]) # if type is not specified correctly for any of these methods, throw an ArgumentError as type is ambiguous in this case.
-            throw(ArgumentError("Unable to infer type automatically from the supplied rule and the supplied type is not supported. Valid types are :irregular and :regular."))
-        end
-    else
-        throw(ArgumentError("The supplied rule, :$rule, is not supported. The rule kwarg must be one of :aic, :bic, :br, :bayes, :mdl, :klcv, :nml, :l2cv, :sturges, :fd, :scott, :wand, :pena, :penb or :penr."))
-    end
-
-    # Fit the histogram here
-    if type == :irregular
-        h = histogram_irregular(x; rule=rule, kwargs...)
-    elseif type == :regular
-        h = histogram_regular(x; rule=rule, kwargs...)
-    end
[…]
