+++ src/AutoHist.jl
@@ -2,8 +2,6 @@ module AutoHist
 
 export AutomaticHistogram
 export fit, convert, loglikelihood, logmarginallikelihood, minimum, maximum, extrema, peaks, pdf, cdf, insupport, length, distance
-export histogram_regular, histogram_irregular
-export DP
 export RRH, Knuth, AIC, BIC, BR, MDL, NML_R, L2CV_R, KLCV_R, Sturges, FD, Scott, Wand
 export RIH, RMG_penA, RMG_penB, RMG_penR, NML_I, L2CV_I, KLCV_I
 export OptPart, SegNeig
@@ -27,17 +25,14 @@ include("utils.jl")
 
 include(joinpath("regular", "objective_functions_regular.jl"))
 include(joinpath("regular", "wand_num_bins.jl"))
+include(joinpath("regular", "regular_rules.jl"))
 
 include(joinpath("irregular", "algorithms.jl"))
 include(joinpath("irregular" ,"greedy_grid.jl"))
 include(joinpath("irregular", "irregular_rules.jl"))
+include(joinpath("irregular", "maximize_additive_crit.jl"))
 include(joinpath("irregular", "objective_functions_irregular.jl"))
 include(joinpath("irregular", "dynprog.jl"))
-include(joinpath("irregular", "compute_bounds.jl"))
-
-include(joinpath("regular", "regular_histogram.jl"))
-include(joinpath("regular", "regular_rules.jl"))
-include(joinpath("irregular", "irregular_histogram.jl"))
 
 function __init__()
     return nothing
