### 2025-07-22 247b235
Updated paper draft examples to use new syntax.
