### 2025-07-22 0c29979
Cleaned up examples a bit.
