+++ previews/PR12/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"api/#API","page":"API","title":"API","text":"","category":"section"},{"location":"api/#The-AutomaticHistogram-type","page":"API","title":"The AutomaticHistogram type","text":"","category":"section"},{"location":"api/#AutoHist.AutomaticHistogram","page":"API","title":"AutoHist.AutomaticHistogram","text":"AutomaticHistogram\n\nA type for representing a histogram where the histogram partition has been chosen automatically based on the sample. Can be fitted to data using the fit method.\n\nFields\n\nbreaks: AbstractVector consisting of the cut points in the chosen partition.\ndensity: Estimated density in each bin.\ncounts: The bin counts for the partition corresponding to breaks.\ntype: Symbol indicating whether the histogram was fit using an irregular procedure (type==:irregular) or a regular one (type==:regular).\nclosed: Symbol indicating whether the drawn intervals should be right-inclusive or not. Possible values are :right (default) and :left.\na: Value of the Dirichlet concentration parameter corresponding to the chosen partition. Only of relevance if a Bayesian method was used to fit the histogram, and is otherwise set to NaN.\n\nExamples\n\njulia> using AutoHist\n\njulia> x = LinRange(eps(), 1.0-eps(), 5000) .^(1.0/4.0);\n\njulia> h = fit(AutomaticHistogram, x)\nAutomaticHistogram\nbreaks: [0.0001220703125, 0.17763663029325183, 0.29718725232110504, 0.4022468898607337, 0.4928155429121377, 0.5797614498414855, 0.6667073567708333, 0.7572760098222373, 0.8405991706295289, 0.9202995853147645, 1.0]\ndensity: [0.006626835974128547, 0.057821970706400425, 0.17596277991076312, 0.36279353706969375, 0.6214544825215076, 0.9730458529384184, 1.4481767793920146, 2.0440057561776532, 2.733509595364622, 3.545742066060377]\ncounts: [5, 34, 92, 164, 270, 423, 656, 852, 1090, 1414]\ntype: irregular\nclosed: right\na: 5.0\n\n\n\n\n\n","category":"type"},{"location":"api/#Fitting-an-automatic-histogram-to-data","page":"API","title":"Fitting an automatic histogram to data","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"An automatic histogram based on regular or irregular partitions can be fitted to the data by calling the fit method.","category":"page"},{"location":"api/#StatsAPI.fit-Tuple{Type{AutomaticHistogram}, AbstractVector{<:Real}, AutoHist.AbstractRule}","page":"API","title":"StatsAPI.fit","text":"fit(AutomaticHistogram, x::AbstractVector{x<:Real}, rule::AbstractRule=RIH(); support::Tuple{Real,Real}=(-Inf,Inf), closed::Symbol=:right)\n\nFit a histogram to a one-dimensional vector x with an automatic and data-based selection of the histogram partition.\n\nArguments\n\nx: 1D vector of data for which a histogram is to be constructed.\n\nKeyword arguments\n\nrule: The criterion used to determine the optimal number of bins. Default value is rule=RIH(), the random irregular histogram.\nclosed: Symbol indicating whether the drawn intervals should be right-inclusive or not. Possible values are :right (default) and :left.\nsupport: Tuple specifying the the support of the histogram estimate. If the first element is -Inf, then minimum(x) is taken as the leftmost cutpoint. Likewise, if the second element is Inf, then the rightmost cutpoint is maximum(x). Default value is (-Inf, Inf), which estimates the support of the data.\n\nReturns\n\nh: An object of type A
[…]
