### 2025-07-22 5509834
Removed some redundant code chunks.
