### 2025-07-22 ac44c09
Cleaned up logmarginallik docstring.
