### 2025-07-23 2f6c3ee
Fixed an error in OptPart docstring and improved the writing in the paper draft.
