+++ paper/paper.md
@@ -10,7 +10,7 @@ authors:
     orcid: 0009-0009-5056-7324
     affiliation: 1
 affiliations:
- - name: Karlsruhe Institute of Technology
+ - name: Scientific Computing Center, Karlsruhe Institute of Technology
    index: 1
 output:
   github_document:
@@ -30,13 +30,13 @@ allowing for simple visualization of the resulting histogram fits.
 To this day, histograms remain one of the most widely used nonparametric density estimators. Their popularity is in no doubt due to their simplicity and interpretability,
 and they are routinely used by practitioners with limited mathematical backgrounds as a means of visualizing the distribution of data sets. 
 Although the construction of a histogram for a given interval partition is a simple task, the quality of the resulting density estimate is very sensitive to the choice of bins.
-As a result, the task of designing automatic histogram procedures, where the number of bins and their location are chosen automatically based on the sample, has recieved considerable interest in the statistics community.
+As a result, the task of designing automatic histogram procedures, where the number of bins and their location are chosen in a data-driven fashion, has recieved considerable interest in the statistics community.
 Despite advances in our understanding of different bin selection rules, many popular software libraries only include a limited number of simple rules for selecting a regular histogram partition,
 and typically do not provide any support for automatic irregular histogram construction.
 `AutoHist.jl` fills this gap by providing a fast implementation of state-of-the-art regular and irregular bin selection algorithms from the statistics literature. A complete overview of the bin selection procedures that have been implemented so far is given in **Table 1**.
 
 | Rule | Type (regular/irregular) | Reference |
-|---------------------|:-----------|:--------------------|
+|--------------------|:-------------------------|:-------------------------|
 | Sturges           | regular   | @sturges1926choice |
 | Scott             | regular   | @scott1979optimal   |
 | Freedman-Diaconis | regular   | @freedman1981histogram |
@@ -45,7 +45,7 @@ and typically do not provide any support for automatic irregular histogram const
 | Birgé-Rozenholc   | regular   | @birge2006bins |
 | MDL               | regular   | @hall1988stochastic |
 | Wand              | regular   | @wand1997data |
-| RRH, Knuth        | regular   | @knuth2019optimal, @simensen2025random |
+| RRH, Knuth        | regular   | @knuth2019optimal |
 | RMG penA          | irregular | @rozenholc2010irregular |
 | RMG penB          | irregular | @rozenholc2010irregular |
 | RMG penR          | irregular | @rozenholc2010irregular |
@@ -57,14 +57,13 @@ and typically do not provide any support for automatic irregular histogram const
 
 Table: Implemented bin selection procedures so far. For methods with type=both, a regular and an irregular variant of the criterion is supported.
 
-
 We note that some automatic histogram selection rules have been implemented in Julia, typically as part of plotting libraries.
-The `Plots.jl` package provides an implementation of some plug-in rules for regular histograms, while the `StatsPlots` package [@christ2023plots] offers an implementation of equal-area histograms.
+The `Plots.jl` package includes support for certain plug-in rules used in standard histograms, whereas the `StatsPlots.jl` library [@christ2023plots] provides functionality for plotting equal-area histograms.
 The `R` package `histogram` [@mildenberger2019histogram] supports some regular and irregular histogram methods, but their implementation covers fewer criteria than ours.
 
 # Installation and usage
 
-The `AutoHist` package is part of the Julia general registry, and can as such be installed via the built in package manager,
+The `AutoHist` package is part of the Julia general registry, and can as such be installed via the built-in package manager,
 ```julia
 using Pkg
 Pkg.add("AutoHist")
[…]
