### 2025-07-23 9c2e22c
Fixed some minor errors in the algorithms documentation.
