### 2025-07-23 0fc053c
Fixed a bug in RMG_penR and started adding doctests to methods.
