+++ src/irregular/irregular_rules.jl
@@ -34,6 +34,24 @@ The default value is ``p_n(k) \\propto 1``. Here, ``a_j = a/k``, for a scalar ``
 - `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)` if `grid` is `regular` or `quantile`. Ignored if `grid=:data`.
 - `alg`: Algorithm used to fit the model. Currently, only `SegNeig()` for this rule. See [`SegNeig`](@ref) for further details.
 
+# Examples
+```jldoctest
+julia> using AutoHist
+
+julia> x = (1.0 .- (1.0 .- LinRange(0.0, 1.0, 500)) .^(1/3)).^(1/3);
+
+julia> rule = RIH(a = 5.0, logprior = k-> -log(k), grid = :data);
+
+julia> fit(AutomaticHistogram, x, rule)
+AutomaticHistogram
+breaks: [0.0, 0.18220071105959446, 0.3587941358096334, 0.8722292888743843, 1.0]
+density: [0.11858322346056327, 0.6490600487586273, 1.6066011289666577, 0.30436411114439915]
+counts: [10, 57, 414, 19]
+type: irregular
+closed: right
+a: 5.0
+```
+
 # References
 This approach to irregular histograms first appeared in [Simensen et al. (2025)](https://doi.org/10.48550/ARXIV.2505.22034).
 """
@@ -80,6 +98,24 @@ Consists of finding the partition ``\\mathcal{I}`` that maximizes a penalized lo
 - `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)` if `grid` is `regular` or `quantile`. Ignored if `grid=:data`.
 - `alg`: Algorithm used to fit the model. Currently, only `SegNeig()` is supported for this rule. See [`SegNeig`](@ref) for further details.
 
+# Examples
+```jldoctest
+julia> using AutoHist
+
+julia> x = (1.0 .- (1.0 .- LinRange(0.0, 1.0, 500)) .^(1/3)).^(1/3);
+
+julia> rule = RMG_penA(grid = :data);
+
+julia> fit(AutomaticHistogram, x, rule)
+AutomaticHistogram
+breaks: [0.0, 0.18875598171056715, 0.3644223879547405, 0.8696799410193466, 1.0]
+density: [0.116552597701164, 0.6717277510418354, 1.6229346697072502, 0.30693663211078037]
+counts: [11, 59, 410, 20]
+type: irregular
+closed: right
+a: NaN
+```
+
 # References
 This approach was suggested by [Rozenholc et al. (2010)](https://doi.org/10.1016/j.csda.2010.04.021).
 """
@@ -122,6 +158,24 @@ Consists of finding the partition ``\\mathcal{I}`` that maximizes a penalized lo
 - `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)` if `grid` is `regular` or `quantile`. Ignored if `grid=:data`.
 - `alg`: Algorithm used to fit the model. Currently, only `SegNeig()` is supported for this rule. See [`SegNeig`](@ref) for further details.
 
+# Examples
+```jldoctest
+julia> using AutoHist
+
+julia> x = (1.0 .- (1.0 .- LinRange(0.0, 1.0, 500)) .^(1/3)).^(1/3);
+
+julia> rule = RMG_penB(grid = :data);
+
+julia> fit(AutomaticHistogram, x, rule)
+AutomaticHistogram
[…]
