### 2025-07-23 643d2cb
Completed doctests on methods page.
