+++ src/regular/regular_rules.jl
@@ -30,6 +30,27 @@ choices ``a_j = 0.5`` and ``p_n(k)\\propto 1``.
 - `logprior`: Unnormalized logprior distribution on the number ``k`` of bins. Defaults to a uniform prior, e.g. `logprior(k) = 0` for all `k`.
 - `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)`.
 
+# Examples
+```jldoctest
+julia> using AutoHist
+
+julia> x = (1.0 .- (1.0 .- LinRange(0.0, 1.0, 500)) .^(1/3)).^(1/3);
+
+julia> rule = RRH(a = k->0.5*k, logprior = k->0.0);
+
+julia> h = fit(AutomaticHistogram, x, rule)
+AutomaticHistogram
+breaks: LinRange{Float64}(0.0, 1.0, 11)
+density: [0.04950495049504951, 0.2079207920792079, 0.5643564356435643, 1.0, 1.495049504950495, 1.8712871287128714, 1.9702970297029703, 1.6732673267326732, 0.9603960396039604, 0.2079207920792079]
+counts: [2, 10, 28, 50, 75, 94, 99, 84, 48, 10]
+type: regular
+closed: right
+a: 5.0
+
+julia> h == fit(AutomaticHistogram, x, Knuth())
+true
+```
+
 # References
 The `Knuth` criterion for histograms was proposed by [Knuth (2019)](https://doi.org/10.1016/j.dsp.2019.102581). The random regular histogram criterion is a generalization.
 """
@@ -99,6 +120,22 @@ where ``n`` is the sample size.
 # Keyword arguments
 - `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)`.
 
+# Examples
+```jldoctest
+julia> using AutoHist
+
+julia> x = (1.0 .- (1.0 .- LinRange(0.0, 1.0, 500)) .^(1/3)).^(1/3);
+
+julia> fit(AutomaticHistogram, x, AIC())
+AutomaticHistogram
+breaks: LinRange{Float64}(0.0, 1.0, 11)
+density: [0.04, 0.2, 0.56, 1.0, 1.5, 1.88, 1.98, 1.68, 0.96, 0.2]
+counts: [2, 10, 28, 50, 75, 94, 99, 84, 48, 10]
+type: regular
+closed: right
+a: NaN
+```
+
 # References
 The aic criterion was proposed by [Taylor (1987)](https://doi.org/10.1093/biomet/74.3.636) for histograms.
 """
@@ -140,6 +177,22 @@ where ``n`` is the sample size.
 
 # Keyword arguments
 - `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)`.
+
+# Examples
+```jldoctest
+julia> using AutoHist
+
+julia> x = (1.0 .- (1.0 .- LinRange(0.0, 1.0, 500)) .^(1/3)).^(1/3);
+
+julia> fit(AutomaticHistogram, x, BIC())
+AutomaticHistogram
+breaks: LinRange{Float64}(0.0, 1.0, 9)
+density: [0.048, 0.336, 0.816, 1.44, 1.904, 1.904, 1.264, 0.288]
+counts: [3, 21, 51, 90, 119, 119, 79, 18]
+type: regular
+closed: right
+a: NaN
+```
 """
[…]
