+++ src/irregular/irregular_rules.jl
@@ -218,6 +218,23 @@ Consists of finding the partition ``\\mathcal{I}`` that maximizes a penalized lo
 - `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)` if `grid` is `regular` or `quantile`. Ignored if `grid=:data`.
 - `alg`: Algorithm used to fit the model. Currently, only `SegNeig()` is supported for this rule. See [`SegNeig`](@ref) for further details.
 
+# Examples
+```jldoctest
+julia> using AutoHist
+
+julia> x = (1.0 .- (1.0 .- LinRange(0.0, 1.0, 500)) .^(1/3)).^(1/3);
+
+julia> rule = RMG_penR(grid = :data);
+
+julia> fit(AutomaticHistogram, x, rule)
+AutomaticHistogram
+breaks: [0.0, 0.18875598171056715, 0.3699070396003733, 0.8285645195146814, 0.9222490305512866, 1.0]
+density: [0.116552597701164, 0.6845115973621804, 1.6875338000474953, 0.7471886144834441, 0.12861575966067232]
+counts: [11, 62, 387, 35, 5]
+type: irregular
+closed: right
+a: NaN
+```
 
 # References
 This approach was suggested by [Rozenholc et al. (2010)](https://doi.org/10.1016/j.csda.2010.04.021).
@@ -328,6 +345,24 @@ Consists of finding the partition ``\\mathcal{I}`` that maximizes a L2 leave-one
 - `alg`: Algorithm used to fit the model. Currently, [`OptPart](@ref) and [`SegNeig`](@ref) are supported for this rule, with the former algorithm being the default.
 - `use_min_length`: Boolean indicating whether or not to impose a restriction on the minimum bin length of the histogram. If set to true, the smallest allowed bin length is set to `(maximum(x)-minimum(x))/n*log(n)^(1.5)`.
 
+# Examples
+```jldoctest
+julia> using AutoHist
+
+julia> x = (1.0 .- (1.0 .- LinRange(0.0, 1.0, 500)) .^(1/3)).^(1/3);
+
+julia> rule = L2CV_I(grid = :data, use_min_length=true);
+
+julia> fit(AutomaticHistogram, x, rule)
+AutomaticHistogram
+breaks: [0.0, 0.149647045210915, 0.2499005080461325, 0.3490626376697454, 0.4600140220788484, 0.7765683248449301, 0.8535131937737716, 0.9121099383916996, 0.9560732934980348, 1.0]
+density: [0.08018868653963065, 0.3590898407087615, 0.7664216197097746, 1.2798398213438569, 1.8448651460332468, 1.2476465466304287, 0.6826317786220794, 0.2729545998246794, 0.045530388214070294]
+counts: [6, 18, 38, 71, 292, 48, 20, 6, 1]
+type: irregular
+closed: right
+a: NaN
+```
+
 # References
 This approach dates back to [Rudemo (1982)](https://www.jstor.org/stable/4615859).
 """
@@ -373,6 +408,24 @@ where the maximmization is over all partitions with ``N_j \\geq 2`` for all ``j`
 - `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)` if `grid` is `regular` or `quantile`. Ignored if `grid=:data`.
 - `alg`: Algorithm used to fit the model. Currently, [`OptPart](@ref) and [`SegNeig`](@ref) are supported for this rule, with the former algorithm being the default.
[…]
