### 2025-07-23 331ad87
Added index to methods page.
