+++ Project.toml
@@ -1,7 +1,7 @@
 name = "AutoHist"
 uuid = "0bf0ee65-3559-4d1e-a282-1b0e88839781"
 authors = ["Oskar Høgberg Simensen"]
-version = "0.7.0"
+version = "0.7.1"
 
 [deps]
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
