### 2025-07-24 f2468af
Irregular histogram rules will now check whether a given algorithm is compatible with the rule. Also fixed some minor doc errors.
