+++ src/irregular/irregular_rules.jl
@@ -32,7 +32,7 @@ The default value is ``p_n(k) \\propto 1``. Here, ``a_j = a/k``, for a scalar ``
 - `logprior`: Unnormalized logprior distribution on the number ``k`` of bins. Defaults to a uniform prior, e.g. `logprior(k) = 0` for all `k`.
 - `grid`: Symbol indicating how the finest possible mesh should be constructed. Options are `:data`, which uses each unique data point as a grid point, `:regular` (default) which constructs a fine regular grid, and `:quantile` which constructs the grid based on the sample quantiles.
 - `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)` if `grid` is `regular` or `quantile`. Ignored if `grid=:data`.
-- `alg`: Algorithm used to fit the model. Currently, only `SegNeig()` for this rule. See [`SegNeig`](@ref) for further details.
+- `alg`: Algorithm used to fit the model. Currently, only [`SegNeig`](@ref) is supported for this rule.
 
 # Examples
 ```jldoctest
@@ -55,12 +55,17 @@ a: 5.0
 # References
 This approach to irregular histograms first appeared in [Simensen et al. (2025)](https://doi.org/10.48550/ARXIV.2505.22034).
 """
-RIH(; a::Real=5.0, logprior::Function=k->0.0, grid::Symbol=:regular, maxbins::Union{Int, Symbol}=:default, alg::AbstractAlgorithm=SegNeig()) = RIH(a, logprior, grid, maxbins, alg)
-
-function fit_autohist(x::AbstractVector{T}, rule::RIH, xmin::T, xmax::T, closed::Symbol) where {T <: Real}
-    if rule.a ≤ 0.0
+function RIH(; a::Real=5.0, logprior::Function=k->0.0, grid::Symbol=:regular, maxbins::Union{Int, Symbol}=:default, alg::AbstractAlgorithm=SegNeig())
+    if typeof(alg) != SegNeig
+        throw(ArgumentError("Algorithm $(typeof(alg)) not supported for rule RIH. Only the SegNeig algorithm is supported for this rule."))
+    end
+    if a ≤ 0.0
         throw(DomainError("Supplied value of a must be positive."))
     end
+    return RIH(a, logprior, grid, maxbins, alg)
+end
+
+function fit_autohist(x::AbstractVector{T}, rule::RIH, xmin::T, xmax::T, closed::Symbol) where {T <: Real}
     n = length(x)
     bin_edges_norm = maximize_additive_crit(x, rule, xmin, xmax, closed)
     
@@ -96,7 +101,7 @@ Consists of finding the partition ``\\mathcal{I}`` that maximizes a penalized lo
 # Keyword arguments
 - `grid`: Symbol indicating how the finest possible mesh should be constructed. Options are `:data`, which uses each unique data point as a grid point, `:regular` (default) which constructs a fine regular grid, and `:quantile` which constructs the grid based on the sample quantiles.
 - `maxbins`: Maximal number of bins for which the above criterion is evaluated. Defaults to `maxbins=:default`, which sets maxbins to the ceil of `min(1000, 4n/log(n)^2)` if `grid` is `regular` or `quantile`. Ignored if `grid=:data`.
-- `alg`: Algorithm used to fit the model. Currently, only `SegNeig()` is supported for this rule. See [`SegNeig`](@ref) for further details.
+- `alg`: Algorithm used to fit the model. Currently, only [`SegNeig`](@ref) is supported for this rule.
 
 # Examples
 ```jldoctest
@@ -119,7 +124,12 @@ a: NaN
[…]
