### 2025-07-24 8a5f78e
Made doctests a bit more compact.
