### 2025-07-25 b33a150
Fixed a bug with maxbins for regular histograms. Moved rule-related error cheks into rule kwarg constructors.
