+++ src/irregular/irregular_rules.jl
@@ -4,7 +4,8 @@ struct RIH{T<:Real, A<:AbstractAlgorithm} <: AbstractIrregularRule
     a::T
     logprior::Function
     grid::Symbol
-    maxbins::Union{Int, Symbol}
+    maxbins::Int
+    use_default_maxbins::Bool
     alg::A
 end
 
@@ -60,7 +61,19 @@ function RIH(; a::Real=5.0, logprior::Function=k->0.0, grid::Symbol=:regular, ma
     if a ≤ 0.0
         throw(DomainError("Supplied value of a must be positive."))
     end
-    return RIH(a, logprior, grid, maxbins, alg)
+    if typeof(maxbins) <: Symbol && maxbins != :default
+        throw(ArgumentError("maxbins must either be a positive integer or :default."))
+    elseif typeof(maxbins) <: Int && maxbins < 1
+        throw(DomainError("maxbins bins must be positive."))
+    end
+    if !(grid in (:data, :regular, :quantile))
+        throw(ArgumentError("Invalid grid option: $(grid). Please select one of :data, :regular or :quantile."))
+    end
+    if maxbins == :default
+        return RIH(a, logprior, grid, 0, true, alg)
+    else
+        return RIH(a, logprior, grid, maxbins, false, alg)
+    end
 end
 
 function fit_autohist(x::AbstractVector{T}, rule::RIH, xmin::T, xmax::T, closed::Symbol) where {T <: Real}
@@ -80,7 +93,8 @@ end
 # RMG penA
 struct RMG_penA{A<:AbstractAlgorithm} <: AbstractIrregularRule
     grid::Symbol
-    maxbins::Union{Int, Symbol}
+    maxbins::Int
+    use_default_maxbins::Bool
     alg::A
 end
 
@@ -124,7 +138,19 @@ function RMG_penA(; grid::Symbol=:regular, maxbins::Union{Int, Symbol}=:default,
     if typeof(alg) != SegNeig
         throw(ArgumentError("Algorithm $(typeof(alg)) not supported for rule RMG_penA. Only the SegNeig algorithm is supported for this rule."))
     end
-    return RMG_penA(grid, maxbins, alg)
+    if typeof(maxbins) <: Symbol && maxbins != :default
+        throw(ArgumentError("maxbins must either be a positive integer or :default."))
+    elseif typeof(maxbins) <: Int && maxbins < 1
+        throw(DomainError("maxbins bins must be positive."))
+    end
+    if !(grid in (:data, :regular, :quantile))
+        throw(ArgumentError("Invalid grid option :$(grid). Please select one of :data, :regular or :quantile."))
+    end
+    if maxbins == :default
+        return RMG_penA(grid, 0, true, alg)
+    else
+        return RMG_penA(grid, maxbins, false, alg)
+    end
 end
 
 function fit_autohist(x::AbstractVector{T}, rule::RMG_penA, xmin::T, xmax::T, closed::Symbol) where {T <: Real}
@@ -143,7 +169,8 @@ end
 # RMG penB
 struct RMG_penB{A<:AbstractAlgorithm} <: AbstractIrregularRule
     grid::Symbol
-    maxbins::Union{Int, Symbol}
+    maxbins::Int
+    use_default_maxbins::Bool
     alg::A
 end
 
@@ -187,7 +214,19 @@ function RMG_penB(; grid::Symbol=:regular, maxbins::Union{Int, Symbol}=:default,
     if typeof(alg) != SegNeig
         throw(ArgumentError("Algorithm $(typeof(alg)) not supported for rule RMG_penB. Only the SegNeig algorithm is supported for this rule."))
     end
-    return RMG_penB(grid, maxbins, alg)
+    if !(grid in (:data, :regular, :quantile))
[…]
