+++ src/regular/regular_rules.jl
@@ -3,7 +3,8 @@
 struct RRH{T} <: AbstractRegularRule
     a::T
     logprior::Function
-    maxbins::Union{Int, Symbol}
+    maxbins::Int
+    use_default_maxbins::Bool
 end
 
 """
@@ -54,11 +55,22 @@ The `Knuth` criterion for histograms was proposed by [Knuth (2019)](https://doi.
 """
 RRH, Knuth
 
-RRH(; a::Union{Real, Function}=5.0, logprior::Function=k->0.0, maxbins::Union{Int, Symbol}=:default) = RRH(a, logprior, maxbins)
+function RRH(; a::Union{Real, Function}=5.0, logprior::Function=k->0.0, maxbins::Union{Int, Symbol}=:default)
+    if typeof(maxbins) <: Symbol && maxbins != :default
+        throw(ArgumentError("maxbins must either be a positive integer or :default."))
+    elseif typeof(maxbins) <: Int && maxbins < 1             # maximal number of bins must be positive
+        throw(DomainError("maxbins bins must be positive."))
+    end
+    if maxbins == :default
+        return RRH(a, logprior, 0, true)
+    else
+        return RRH(a, logprior, maxbins, false)
+    end
+end
 
 function fit_autohist(x::AbstractVector{T}, rule::RRH, xmin::T, xmax::T, closed::Symbol) where {T <: Real}
     n = length(x)
-    k_max = get_maxbins_regular(rule.maxbins, n)
+    k_max = ifelse(rule.use_default_maxbins, min(ceil(Int, 4.0*n / log(n)^2), 1000), rule.maxbins)
     a_vec = Vector{Float64}(undef, k_max)
     if !isa(rule.a, Function) # create constant function if typeof(a) <: Real
         if rule.a ≤ 0.0
@@ -92,16 +104,14 @@ end
 
 # ------------------------------
 # Knuth's rule
-struct Knuth <: AbstractRegularRule
-    maxbins::Union{Int, Symbol}
-end
-Knuth(; maxbins::Union{Int, Symbol}=:default) = RRH(k->0.5*k, k->0.0, maxbins)
+Knuth(; maxbins::Union{Int, Symbol}=:default) = RRH(; a=k->0.5*k, logprior=k->0.0, maxbins=maxbins)
 
 
 # ------------------------------
 # AIC histogram
 struct AIC <: AbstractRegularRule
-    maxbins::Union{Int, Symbol}
+    maxbins::Int
+    use_default_maxbins::Bool
 end
 
 """
@@ -135,11 +145,22 @@ a: NaN
 # References
 The aic criterion was proposed by [Taylor (1987)](https://doi.org/10.1093/biomet/74.3.636) for histograms.
 """
-AIC(; maxbins::Union{Int, Symbol}=:default) = AIC(maxbins)
+function AIC(; maxbins::Union{Int, Symbol}=:default)
+    if typeof(maxbins) <: Symbol && maxbins != :default
+        throw(ArgumentError("maxbins must either be a positive integer or :default."))
+    elseif typeof(maxbins) <: Int && maxbins < 1             # maximal number of bins must be positive
+        throw(DomainError("maxbins bins must be positive."))
+    end
+    if maxbins == :default
+        return AIC(0, true)
+    else
+        return AIC(maxbins, false)
+    end
+end
 
 function fit_autohist(x::AbstractVector{T}, rule::AIC, xmin::T, xmax::T, closed::Symbol) where {T <: Real}
     n = length(x)
-    k_max = get_maxbins_regular(rule.maxbins, n)
+    k_max = ifelse(rule.use_default_maxbins, min(ceil(Int, 4.0*n / log(n)^2), 1000), rule.maxbins)
[…]
