### 2025-07-25 91d2fb4
Added a test for maxbins != default.
