### 2025-07-29 1f56744
Fixed logmarginallik docstring.
