### 2025-07-30 efc8d96
Removed missing reference from docs.
