+++ paper/paper.bib
@@ -12,10 +12,9 @@ doi = {10.1137/141000671},
 @article{christ2023plots,
   doi = {10.5334/jors.431},
   author = {Christ, Simon and Schwabeneder, Daniel and Rackauckas, Christopher and Borregaard, Michael Krabbe and Breloff, Thomas},
-  title = {Plots.jl -- a user extendable plotting API for the julia programming language},
+  title = {Plots.jl - a user extendable plotting API for the julia programming language},
   publisher = {Journal of Open Research Software},
-  year = {2023},
-  copyright = {Creative Commons Attribution 4.0 International}
+  year = {2023}
 }
 
 @article{danisch2021makie,
@@ -25,7 +24,7 @@ doi = {10.1137/141000671},
   volume = {6},
   pages = {3349},
   author = {Danisch, Simon and Krumbiegel, Julius},
-  title = {Makie.jl: Flexible high-performance data visualization for Julia},
+  title = {Makie.jl: Flexible high-performance data visualization for {J}ulia},
   journal = {Journal of Open Source Software}
   }
 
@@ -82,7 +81,8 @@ doi = {10.1137/141000671},
   journal={Scandinavian Journal of Statistics},
   pages={65--78},
   year={1982},
-  volume={9}
+  volume={9},
+  url={https://www.jstor.org/stable/4615859}
 }
 
 @article{hall1988stochastic,
@@ -117,7 +117,7 @@ doi = {10.1137/141000671},
 
 
 @article{kontkanen2007mdl,
-  title = {MDL Histogram Density Estimation},
+  title = {{MDL} Histogram Density Estimation},
   author = {Kontkanen, Petri and Myllymäki, Petri},
   pages = {219--226},
   year = {2007},
@@ -137,7 +137,7 @@ doi = {10.1137/141000671},
 }
 
 @article{freedman1981histogram,
-  author = {Freedman, D., Diaconis, P.},
+  author = {Freedman, D. and Diaconis, P.},
   title = {On the histogram as a density estimator: L2 theory},
   journal = {Zeitschrift für Wahrscheinlichkeitstheorie und Verwandte Gebiete},
   volume = {57},
