### 2025-08-06 #8615 @jbytecode
@oskarhs - Hi, thank you for submitting your work to JOSS. I'm the handling editor of your submission.

Could you please

- Fix the DOI issue as indicated above posts
- Use `@editorialbot check references` to ensure we get things correct.

Do you have any potential reviewers? You can use the [List of Peope](https://reviewers.joss.theoj.org/lookup) to search and filter potential reviewers. Please provide me a list of gitgub handles. Please do not use `@` character to avoid unwanted notifications.

If you don't have any suggestions, I'll try to find and assign reviewers.

Thank you in advance.
