### 2025-08-08 4dfaaca
Fixed some minor typos/errors in the documentation.
