### 2025-08-10 #8615 @jbytecode
👋👋👋 Dear @everval 👋👋👋

**[AutoHist.jl: A Julia package for fast and automatic histogram construction](https://github.com/openjournals/joss-reviews/issues/8615)**

You can find more information at the top of this Github issue ([LINK](https://github.com/openjournals/joss-reviews/issues/8615)).

Thank you in advance!
