### 2025-08-17 c6fae75
Cleaned up method docstrings.
