### 2025-08-22 #8850 @editorialbot
[REVIEW]: AutoHist.jl: A Julia package for fast and automatic histogram construction
**Submitting author:** @oskarhs (<a href="http://orcid.org/0009-0009-5056-7324">Oskar Høgberg Simensen</a>)
**Repository:** https://github.com/oskarhs/AutoHist.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v0.6.0
