### 2025-08-22 dd1d5e0
Changed a few formulations in the paper.
