### 2025-08-22 #8850 @ajinkya-k
@jbytecode i had a question: the syntax used in the submission PDF doesnt work and it differs from what they have in their github readme. should i still make an issue on their repo for that? basically they need to update the manuscript with the newer API
