### 2025-08-29 6285d37
Added full names of all method to paper.md
