### 2025-08-29 f478fc1
Added quantile method for AutomaticHistogram.
