+++ src/AutomaticHistogram.jl
@@ -135,6 +135,13 @@ function fit(::Type{AutomaticHistogram}, x::AbstractVector{<:Real}, rule::Abstra
 end
 
 """
+    autohist(x::AbstractVector{<:Real}, rule::AbstractRule=RIH(); support::Tuple{Real,Real}=(-Inf,Inf), closed::Symbol=:right)
+
+Fit an automatic histogram to data based on the supplied rule. This is an alias for `fit(AutomaticHistogram, x, rule; support, closed)`. See [`fit`](@ref) for further details.
+"""
+autohist(x::AbstractVector{<:Real}, rule::AbstractRule=RIH(); support::Tuple{Real,Real}=(-Inf,Inf), closed::Symbol=:right) = fit(AutomaticHistogram, x, rule; support, closed)
+
+"""
     convert(Histogram, h::AutomaticHistogram)
 
 Convert an `h` to a StatsBase.Histogram, normalized to be a probability density.
@@ -230,20 +237,20 @@ function pdf(h::AutomaticHistogram, x::Real)
         if h.type == :irregular
             if h.closed == :right
                 idx = max(1, searchsortedfirst(h.breaks, x) - 1)
-                @inbounds val = h.density[idx]
+                val = h.density[idx]
             else 
                 idx = min(k, searchsortedlast(h.breaks, x))
-                @inbounds val = h.density[idx]
+                val = h.density[idx]
             end
         else
             xmin, xmax = extrema(h)
             edges_inc = k/(xmax-xmin)
             if h.closed == :right
                 idx = max(0, floor(Int, (x-xmin)*edges_inc-10.0*edges_inc*eps())) + 1
-                @inbounds val = h.density[idx]
+                val = h.density[idx]
             else 
                 idx = min(k-1, floor(Int, (x-xmin)*edges_inc+10.0*edges_inc*eps())) + 1
-                @inbounds val = h.density[idx]
+                val = h.density[idx]
             end
         end
     end
@@ -272,19 +279,19 @@ function cdf(h::AutomaticHistogram, x::Real)
         if h.type == :irregular
             if h.closed == :right
                 idx = max(1, searchsortedfirst(h.breaks, x) - 1)
-                @inbounds val = h.density[idx] * (x - h.breaks[idx]) + sum(diff(h.breaks[1:idx]) .* h.density[1:idx-1])
+                val = h.density[idx] * (x - h.breaks[idx]) + sum(diff(h.breaks[1:idx]) .* h.density[1:idx-1])
             else 
                 idx = min(k, searchsortedlast(h.breaks, x))
-                @inbounds val = h.density[idx] * (x - h.breaks[idx]) + sum(diff(h.breaks[1:idx]) .* h.density[1:idx-1])
+                val = h.density[idx] * (x - h.breaks[idx]) + sum(diff(h.breaks[1:idx]) .* h.density[1:idx-1])
             end
         else
             edges_inc = k/(xmax-xmin)
             if h.closed == :right
                 idx = max(0, floor(Int, (x-xmin)*edges_inc-10.0*edges_inc*eps())) + 1
-                @inbounds val = h.density[idx] * (x - h.breaks[idx]) + sum(diff(h.breaks[1:idx]) .* h.density[1:idx-1])
+                val = h.density[idx] * (x - h.breaks[idx]) + sum(diff(h.breaks[1:idx]) .* h.density[1:idx-1])
             else 
[…]
