### 2025-08-29 4a43840
Added the Bayesian Blocks method.
