+++ paper/paper.bib
@@ -163,3 +163,13 @@ doi = {10.1137/141000671},
   year={2025},
   doi={10.48550/arXiv.2505.22034}
 }
+
+@article{scargle2013bayesblocks,
+  title = {Studies in astronomical time series analysis. {VI}. {B}ayesian block representations},
+  author = {Scargle, J.~D. and Norris, J.~P. and Jackson, B. and Chiang, J.},
+  journal = {The Astrophysical Journal},
+  volume = {764},
+  pages = {1--26},
+  year = {2013},
+  doi={10.1088/0004-637X/764/2/167}
+}
