+++ src/irregular/irregular_rules.jl
@@ -11,11 +11,11 @@ end
 
 """
     RIH(;
-        a::Real                     = 5.0,
-        logprior::Function          = k -> 0.0,
-        grid::Symbol                = :regular,
-        maxbins::Union{Int, Symbol} = :default,
-        alg::AbstractAlgorithm      = SegNeig()
+        a::Real                    = 5.0,
+        logprior::Function         = k -> 0.0,
+        grid::Symbol               = :regular,
+        maxbins::Union{Int,Symbol} = :default,
+        alg::AbstractAlgorithm     = SegNeig()
     )
 
 The random irregular histogram criterion.
@@ -54,7 +54,7 @@ a: 5.0
 # References
 This approach to irregular histograms first appeared in [Simensen et al. (2025)](https://doi.org/10.48550/ARXIV.2505.22034).
 """
-function RIH(; a::Real=5.0, logprior::Function=k->0.0, grid::Symbol=:regular, maxbins::Union{Int, Symbol}=:default, alg::AbstractAlgorithm=SegNeig())
+function RIH(; a::Real=5.0, logprior::Function=k->0.0, grid::Symbol=:regular, maxbins::Union{Int,Symbol}=:default, alg::AbstractAlgorithm=SegNeig())
     if typeof(alg) != SegNeig
         throw(ArgumentError("Algorithm $(typeof(alg)) not supported for rule RIH. Only the SegNeig algorithm is supported for this rule."))
     end
@@ -64,7 +64,7 @@ function RIH(; a::Real=5.0, logprior::Function=k->0.0, grid::Symbol=:regular, ma
     if typeof(maxbins) <: Symbol && maxbins != :default
         throw(ArgumentError("maxbins must either be a positive integer or :default."))
     elseif typeof(maxbins) <: Int && maxbins < 1
-        throw(DomainError("maxbins bins must be positive."))
+        throw(DomainError(maxbins, "maxbins bins must be positive."))
     end
     if !(grid in (:data, :regular, :quantile))
         throw(ArgumentError("Invalid grid option: $(grid). Please select one of :data, :regular or :quantile."))
@@ -100,9 +100,9 @@ end
 
 """
     RMG_penA(;
-        grid::Symbol                = :regular,
-        maxbins::Union{Int, Symbol} = :default,
-        alg::AbstractAlgorithm      = SegNeig()
+        grid::Symbol               = :regular,
+        maxbins::Union{Int,Symbol} = :default,
+        alg::AbstractAlgorithm     = SegNeig()
     )
 
 Consists of finding the partition ``\\mathcal{I}`` that maximizes a penalized log-likelihood,
@@ -134,14 +134,14 @@ a: NaN
 # References
 This approach was suggested by [Rozenholc et al. (2010)](https://doi.org/10.1016/j.csda.2010.04.021).
 """
-function RMG_penA(; grid::Symbol=:regular, maxbins::Union{Int, Symbol}=:default, alg::AbstractAlgorithm=SegNeig())
+function RMG_penA(; grid::Symbol=:regular, maxbins::Union{Int,Symbol}=:default, alg::AbstractAlgorithm=SegNeig())
     if typeof(alg) != SegNeig
         throw(ArgumentError("Algorithm $(typeof(alg)) not supported for rule RMG_penA. Only the SegNeig algorithm is supported for this rule."))
     end
     if typeof(maxbins) <: Symbol && maxbins != :default
         throw(ArgumentError("maxbins must either be a positive integer or :default."))
     elseif typeof(maxbins) <: Int && maxbins < 1
-        throw(DomainError("maxbins bins must be positive."))
[…]
