### 2025-08-29 fef0e4e
Changed AutomaticHistogram to use type parameters. Also fixed some errors in the docs.
