+++ src/regular/regular_rules.jl
@@ -1,15 +1,15 @@
 # ------------------------------
 # Random regular histogram 
-struct RRH{T} <: AbstractRegularRule
+struct RRH{T, F<:Function} <: AbstractRegularRule
     a::T
-    logprior::Function
+    logprior::F
     maxbins::Int
     use_default_maxbins::Bool
 end
 
 """
     RRH(;
-        a::Union{Real, Function}   = 5.0,
+        a::Union{Real,Function}    = 5.0,
         logprior::Function         = k->0.0,
         maxbins::Union{Int,Symbol} = :default
     )
@@ -42,7 +42,7 @@ julia> x = (1.0 .- (1.0 .- LinRange(0.0, 1.0, 500)) .^(1/3)).^(1/3);
 julia> rule = RRH(a = k->0.5*k, logprior = k->0.0);
 
 julia> h = fit(AutomaticHistogram, x, rule)
-AutomaticHistogram
+AutomaticHistogram{LinRange{Float64, Int64}, Vector{Float64}, Vector{Int64}}
 breaks: LinRange{Float64}(0.0, 1.0, 11)
 density: [0.04950495049504951, 0.2079207920792079, 0.5643564356435643, 1.0, 1.495049504950495, 1.8712871287128714, 1.9702970297029703, 1.6732673267326732, 0.9603960396039604, 0.2079207920792079]
 counts: [2, 10, 28, 50, 75, 94, 99, 84, 48, 10]
@@ -59,7 +59,7 @@ The `Knuth` criterion for histograms was proposed by [Knuth (2019)](https://doi.
 """
 RRH, Knuth
 
-function RRH(; a::Union{Real, Function}=5.0, logprior::Function=k->0.0, maxbins::Union{Int, Symbol}=:default)
+function RRH(; a::Union{Real,Function}=5.0, logprior::Function=k->0.0, maxbins::Union{Int,Symbol}=:default)
     if typeof(maxbins) <: Symbol && maxbins != :default
         throw(ArgumentError("maxbins must either be a positive integer or :default."))
     elseif typeof(maxbins) <: Int && maxbins < 1             # maximal number of bins must be positive
@@ -108,7 +108,7 @@ end
 
 # ------------------------------
 # Knuth's rule
-Knuth(; maxbins::Union{Int, Symbol}=:default) = RRH(; a=k->0.5*k, logprior=k->0.0, maxbins=maxbins)
+Knuth(; maxbins::Union{Int,Symbol}=:default) = RRH(; a=k->0.5*k, logprior=k->0.0, maxbins=maxbins)
 
 
 # ------------------------------
@@ -137,7 +137,7 @@ where ``n`` is the sample size.
 julia> x = (1.0 .- (1.0 .- LinRange(0.0, 1.0, 500)) .^(1/3)).^(1/3);
 
 julia> fit(AutomaticHistogram, x, AIC())
-AutomaticHistogram
+AutomaticHistogram{LinRange{Float64, Int64}, Vector{Float64}, Vector{Int64}}
 breaks: LinRange{Float64}(0.0, 1.0, 11)
 density: [0.04, 0.2, 0.56, 1.0, 1.5, 1.88, 1.98, 1.68, 0.96, 0.2]
 counts: [2, 10, 28, 50, 75, 94, 99, 84, 48, 10]
[…]
