+++ ext/PlotsExt.jl
@@ -5,6 +5,6 @@ import AutoHist: AutomaticHistogram
 import StatsBase: Histogram
 
 # Plot the corresponding StatsBase.Histogram
-@recipe f(::Type{AutomaticHistogram}, h::AutomaticHistogram) = convert(Histogram, h)
+@recipe f(h::AutomaticHistogram) = convert(Histogram, h)
 
 end
