+++ docs/src/examples/algorithm_choice.md
@@ -68,7 +68,7 @@ BenchmarkTools.Trial: 5 samples with 1 evaluation per sample.
 The above code snippet has a median runtime of about ``1.2\ \text{s}`` on my machine.
 
 ## The quadratic-time dynamic programming algorithm
-For the [`L2CV_I`](@ref) and [`KLCV_I`](@ref) criteria, it becomes possible to compute the exact solution via the quadratic-time dynamic programming algorithm [`OptPart`](@ref) instead of the cubic-time `SegNeig` algorithm used for the other problems. In practice, this means that computing the exact solution is often feasible even as the number of candidate cutpoints becomes quite large. Both the `OptPart` and `SegNeig` algorithms can be used to fit `L2CV_I`, allowing us to get a direct comparison between the performance of the two algorithms.
+For the [`L2CV_I`](@ref), [`KLCV_I`](@ref) and [`BayesBlocks`](@ref) criteria, it becomes possible to compute the exact solution via the quadratic-time dynamic programming algorithm [`OptPart`](@ref) instead of the cubic-time `SegNeig` algorithm used for the other problems. In practice, this means that computing the exact solution is often feasible even as the number of candidate cutpoints becomes quite large. Both the `OptPart` and `SegNeig` algorithms can be used to fit `L2CV_I`, allowing us to get a direct comparison between the performance of the two algorithms.
 ```julia
 n = 10^3
 @benchmark fit(
@@ -106,4 +106,4 @@ Due to the superior runtime complexity of the exact algorithm for cross-validati
 
 
 !!! note
-    In practice one should of course never use the `SegNeig` algorithm for the two cross-validation criteria, as it only leads to a reduction in speed compared to using `OptPart`, which is the default.
+    In practice one should of course never use the `SegNeig` algorithm for the `L2CV_I`, `KLCV_I` and `BayesBlocks` criteria, as it only leads to a worse speed/accuracy tradeoff compared to using `OptPart`, which is the default.

[…]
