### 2025-08-29 #8850 @ajinkya-k
@oskarhs Thanks for the quick update. Two suggestions:

1. Since `Plots.jl` is mentioned in the paper could you also add example usage for that in the paper? The one in the readme should suffice.
2. You point out in the paper that other software packages only implement some of the rules, could you give some examples? For instance you could mention what rules `ggplot2` and base R `hist` implement and which they don’t. Similarly for a few of the commonly used python and Julia packages. Just a few sentences should suffice.
