### 2025-09-01 d22a72f
Implemented proposed changes to the paper draft.
