+++ paper/paper.bib
@@ -173,3 +173,23 @@ doi = {10.1137/141000671},
   year = {2013},
   doi={10.1088/0004-637X/764/2/167}
 }
+
+@Manual{Rlanguage2025,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2025},
+  url = {https://www.R-project.org/},
+  note = {version 4.5.1}
+}
+
+@Article{hunter2007matplotlib,
+  author    = {Hunter, J. D.},
+  title     = {Matplotlib: A 2D graphics environment},
+  journal   = {Computing in Science \& Engineering},
+  volume    = {9},
+  pages     = {90--95},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = {2007}
+}
