+++ README.md
@@ -1,7 +1,9 @@
 # AutoHist.jl
 
-[![Build Status](https://github.com/oskarhs/AutoHist.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/oskarhs/AutoHist.jl/actions/workflows/CI.yml?query=branch%3Amain) [![codecov](https://codecov.io/gh/oskarhs/AutoHist.jl/branch/main/graph/badge.svg)](https://codecov.io/gh/oskarhs/AutoHist.jl)
-
+[![Build Status](https://github.com/oskarhs/AutoHist.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/oskarhs/AutoHist.jl/actions/workflows/CI.yml?query=branch%3Amain)
+[![codecov](https://codecov.io/gh/oskarhs/AutoHist.jl/branch/main/graph/badge.svg)](https://codecov.io/gh/oskarhs/AutoHist.jl)
+[![Doc](https://img.shields.io/badge/docs-stable-blue.svg)](https://oskarhs.github.io/AutoHist.jl/stable/)
+[![Doc](https://img.shields.io/badge/docs-dev-blue.svg)](https://oskarhs.github.io/AutoHist.jl/dev/)
 
 A pure Julia implementation of automatic regular and irregular histogram methods based on maximizing a goodness-of-fit criterion.
 
