### 2025-09-01 #8850 @oskarhs
@ajinkya-k

Thank you for the feedback you have provided so far!

I have made some changes to the paper draft in accordance with your suggestions:

1) I have chosen to showcase `Plots.plot()` as a brief "inline" example.
2) A brief overview of automatic binning procedures provided by Plots.jl, base R and matplotlib is now provided. I have opted not to list all the automatic binning rules implemented by matplotlib, as some of these, to my knowledge, lack proper theoretical justification (such as the square root or Rice rules). Listing all of the methods from the histogram R package would (IMO) take up too much space, and I have therefore omitted doing this.

I look forward to hearing more from you!
