### 2025-09-16 oskarhs/AutoHist.jl#14 @PieterjanRobbe
Documentation > Density estimation
When installing the `TestDistributions` package as detailed in the [documentation](https://oskarhs.github.io/AutoHist.jl/stable/examples/density_estimation/#Mode-hunting), I believe you need to say

```julia
import TestDistributions as TD
```

instead of

```julia
import .TestDistributions as TD
```

[breadcrumb](https://github.com/openjournals/joss-reviews/issues/8850)
