### 2025-09-16 oskarhs/AutoHist.jl#15 @PieterjanRobbe
Custom formatting of histograms
I think it would be nice if there were some examples showing how you can format the histograms plotted with `Plots` or `Makie`. Judging by the plots in the documentation, this seems possible, but I couldn't immediately figure out how to do this.

[breadcrumb](https://github.com/openjournals/joss-reviews/issues/8850)
