### 2025-09-16 oskarhs/AutoHist.jl#16 @PieterjanRobbe
Difference with arXiv paper?
It seems like [this](https://arxiv.org/pdf/2505.22034) arXiv paper already mentions `AutoHist.jl`. Is there anything that is novel compared to the version used for the arXiv paper? Depending on which journal you submitted this article to, you will need to upload your data and source code. Will the JOSS paper create a duplicate entry in this case?

[breadcrumb](https://github.com/openjournals/joss-reviews/issues/8850)
