### 2025-09-16 oskarhs/AutoHist.jl#17 @PieterjanRobbe
Table of Contents
The Table of Contents in the documentation says "Introdution" instead of "Introduction".

[breadcrumb](https://github.com/openjournals/joss-reviews/issues/8850)
