### 2025-09-16 oskarhs/AutoHist.jl#20 @PieterjanRobbe
JOSS paper
line 45: "aabove" -> "above"

[breadcrumb](https://github.com/openjournals/joss-reviews/issues/8850)
