### 2025-09-16 39f9a4d
Added functionality for direct plotting of automatic histograms. Also wrote a short plotting tutorial.
