### 2025-09-16 05ed50f
Added compat entry for CairoMakie
