### 2025-09-17 516927a
Fixed plotting with Makie 0.24
