+++ versions.js
@@ -9,5 +9,5 @@ var DOC_VERSIONS = [
   "v0.2",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.0.0";
+var DOCUMENTER_NEWEST = "v1.0.1";
 var DOCUMENTER_STABLE = "stable";
