### 2025-09-17 oskarhs/AutoHist.jl#20 @oskarhs
Corrected in #21
