### 2025-09-17 047eb4b
Added logo to Readme
