### 2025-09-17 f912ff9
Fixed logo not rendering.
