### 2025-09-17 59f4131
Added darkmode logo.
