### 2025-09-17 8e2810b
Did some logo formatting.
