### 2025-09-17 dde12d3
Added a footnote on supported seriestypes for Plots.jl
