+++ docs/src/algorithms.md
@@ -7,7 +7,7 @@ All the irregular histogram methods supported by this library are the product of
 ```math
     \max_{1\leq k\leq k_n}\max_{\boldsymbol{t}_{0:k}} \Big\{\sum_{j=1}^k \Phi(\tau_{n, t_{j-1}}, \tau_{n, t_{j}}] + \Psi(k)\Big\},
 ```
-where the inner maximum is over integer vectors ``\boldsymbol{t}_{0:k}`` satisfying ``0 = t_0 < t_1 < \cdots < t_{k-1} < t_k = k_n`` and ``\{\tau_{n,j}\colon 0\leq j \leq k_n\}`` are the candidate cutpoints between the partition intervals. While it is generally desirable to use a moderate number of possible cutpoints relative to the number of samples, this comes at a heavy computational cost if an exact solution of the above optimization is desired as the runtime complexity of the algorithm is cubic in the number of candidates. We note that for the special case where ``\Psi(k) = 0`` for all ``k``, a more efficient algorithm is available, allowing for a quadratic-time solution. In both cases, computing the exact solution becomes unfeasible for larger sample sizes.
+where the inner maximum is over integer vectors ``\boldsymbol{t}_{0:k}`` satisfying ``0 = t_0 < t_1 < \cdots < t_{k-1} < t_k = k_n`` and ``\{\tau_{n,j}\colon 0\leq j \leq k_n\}`` are the candidate cutpoints between the partition intervals. While it is generally desirable to use a moderate number of possible cutpoints relative to the number of samples, this comes at a heavy computational cost if an exact solution of the above optimization is desired as the runtime complexity of the algorithm is cubic in the number of candidates. We note that for the special case where ``\Psi(k) = \beta k`` for some scalar ``\beta\in \mathbb{R}`` and all ``k``, a more efficient algorithm is available, allowing for a quadratic-time solution. In both cases, computing the exact solution becomes unfeasible for larger sample sizes.
 
 To ease the computational burden we adopt a greedy search heuristic to construct a subset of possible cutpoints when the number of candidates is large, and then run the optimization algorithm of choice on this smaller set. To find the optimal partition for a given set of candidate cutpoints, this package includes two solvers based on dynamic programming. The algorithm used can be controlled via the `alg` keyword passed to the rule argument of [`fit`](@ref), see the [methods page](methods.md#irregular-histograms).
 
