### 2025-09-17 #8850 @PieterjanRobbe
Hi @oskarhs, thanks for the quick follow-up! Looks great to me.

@jbytecode: I've finished my review.
