+++ README.md
@@ -1,17 +1,5 @@
-<table style="border-collapse: collapse; border: none;">
-  <tr>
-    <td valign="middle" style="border: none; padding: 0;">
-      <picture>
-        <source srcset="docs/src/assets/logo-dark.svg" media="(prefers-color-scheme: dark)">
-        <img src="docs/src/assets/logo.svg" width="90" alt="AutoHist logo">
-      </picture>
-    </td>
-    <td valign="middle" style="border: none; padding-left: 10px;">
-      <h1 style="margin: 0;">AutoHist.jl</h1>
-    </td>
-  </tr>
-</table>
-
+<picture><source srcset="docs/src/assets/logo-dark.svg" media="(prefers-color-scheme: dark)"><img src="docs/src/assets/logo.svg" width="90" alt="AutoHist logo" align="left"></picture>
+# AutoHist.jl <br><br>
 
 [![Build Status](https://github.com/oskarhs/AutoHist.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/oskarhs/AutoHist.jl/actions/workflows/CI.yml?query=branch%3Amain)
 [![codecov](https://codecov.io/gh/oskarhs/AutoHist.jl/branch/main/graph/badge.svg)](https://codecov.io/gh/oskarhs/AutoHist.jl)
