### 2025-09-18 b0a6991
Simplified the plotting examples on the index page.
