+++ docs/src/examples/plotting.md
@@ -1,6 +1,6 @@
 # Plotting tutorial
 
-The following markdown page gives a brief, example-driven introduction to plotting histograms with AutoHist.jl. All the plotting functionality is provided through extensions for Makie.jl and Plots.jl. Both extensions provide recipes for plotting [`AutomaticHistogram`](@ref) objects, and additionally lets the user utilize the automatic bin selection rules from AutoHist when calling the respective built-in histogram plotting functions from the two libraries.
+The following markdown page gives a brief, example-driven introduction to plotting histograms with AutoHist.jl. All the plotting functionality is provided through extensions for [Makie.jl](https://docs.makie.org/stable/) and [Plots.jl](https://docs.juliaplots.org/stable/). Both extensions provide recipes for plotting [`AutomaticHistogram`](@ref) objects, and additionally lets the user utilize the automatic bin selection rules from AutoHist when calling the respective built-in histogram plotting functions from the two libraries.
 
 ## Plots.jl
 In order to use the AutoHist.jl plotting functionality with Plots.jl, we will need to load both libraries first.
@@ -35,7 +35,7 @@ As illustrated here, the call to `Plots.plot` also accepts additional keyword ar
     In particular, the seriestypes `histogram`, `barbins`, `barhist` and `bar` will draw the histogram as a filled bar chart, while `stephist` draws a non-filled histogram. An attempt to plot an `AutomaticHistogram` with another [seriestype](https://docs.juliaplots.org/latest/generated/attributes_series/) throws an error.
 
 !!! note
-    Some of the arguments normally supported by the `histogram` and `stephist` seriestypes are deprecated when used together with `AutomaticHistogram`. In particular, the `weights` keyword is not supported, as `autohist` currently only handles equally weighted samples. The `bins` keyword is ignored, as the histogram partition has already been chosen by `autohist`. Additionally, the `normalize` kwarg is deprecated, so that e.g. `Plots.histogram(h, normalize=norm)` will always draw a histogram on the density scale no matter the value of `norm`.
+    Some of the arguments normally supported by the `histogram` and `stephist` seriestypes are deprecated when used together with `AutomaticHistogram`. In particular, the `weights` keyword is not supported, as `fit`/`autohist` currently only handles equally weighted samples. The `bins` keyword is ignored, as the histogram partition has already been chosen by `autohist`. Additionally, the `normalize` kwarg is deprecated, so that e.g. `Plots.histogram(h, normalize=norm)` will always draw a histogram on the density scale no matter the value of `norm`.
 
 Automatic histograms can also be plotted by accessing Plots.jl functions directly. The general syntax for fitting a histogram in this manner is `Plots.plot(x, rule)`, where `rule` is any of the bin selection rules provided by the AutoHist.jl package, see the [methods page](../methods.md). Note that the above syntax supports any histogram-like seriestype, so the above inline code snippet is equivalent to `Plots.histogram(x, rule)`.
 When using Plots.jl functions directly, one can pass additional keyword arguments directly to the plotting function. In the code snippet below, we show how to directly plot a right-closed histogram with support constrained to be positive.
@@ -89,7 +89,7 @@ fig # Display figure
 As shown in the above example, we can pass additional plot attributes to the Makie plotting functions in order to customize our histogram plots. We refer the interested reader to the [Makie documentation](https://docs.makie.org/stable/reference/plots/hist) for further details on the supported attributes.
 
 !!! note
[…]
