### 2025-09-18 15c17d7
add doi to citation of R
