+++ paper/paper.bib
@@ -181,7 +181,8 @@ doi = {10.1137/141000671},
   address = {Vienna, Austria},
   year = {2025},
   url = {https://www.R-project.org/},
-  note = {version 4.5.1}
+  note = {version 4.5.1},
+  doi = {10.32614/r.manuals}
 }
 
 @Article{hunter2007matplotlib,
