### 2025-09-18 #8850 @jbytecode
@oskarhs - A new PR (https://github.com/oskarhs/AutoHist.jl/pull/23), could you please merge?
